Moved secured to package essix/secure/db
As to ensure its initialization before secure.Configure()
